Show users requested raw permissions or an error
Add base64 url decoding in Elm and pass the decoded JSON string to JavaScript.
Decoding lets us show the user the requested resources as JSON, and let's us
remove decoding in JavaScript.
On failures to decode the query param or parse the JSON, we show an error
informing the user the request is invalid and will be ignored, and we pass
JavaScript an empty list (as a JSON string).