Fix apps concurrently trying to authenticate
Originally I thought this was just too much of an edge case to support.
And I weighted this against "exposing as little encrypted data as possible".
But I think that's not a valid concern.

Overriding the data and removing the data leads to far more difficult
bugs and edge cases to handle in webnative than the safety it privides (which is basically 'none').