Add initial ION DID interop implementation
Adds a demohacked sign up page for users to sign up using their ION DID
and private key. This should not be merged! No guarantees are made about
security and it changes the sign up page in bold ways that should be more
carefuly considered.
The current implemenation goes as far as validating the DID is on the ION
network and preparing the UCAN and JWT. It does not call the API to create the
account yet.